16: adding in checks for transient dummy vm.
Adding in a more thorough check for transient VM's. If the OVMF shell vm
is found to be running after waiting an acceptable alotted time,
virDomainDestory command is given to destroy the transient VM. This
should.

Also added in some timing logic to help reduce the margin of error for
the OVMF shell vm starting and shutting down properly.

Signed-off-by: Larry Dewey <ldewey@suse.com>