Fix misplaced bracket in the sev::read_file() of tmp_tk.bin
The expression has the third argument as

sizeof(m_tk) != sizeof(m_tk)

which always returns zero meaning the code does a zero length read of
tmp_tk.bin and thus m_tk ends up containing stack junk which causes
sev-tool to build an incorrect secret package.

I think the intention is for the third argument to be sizeof(m_tk) and
then for the return to be checked against this value.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>