Use the system supplied source of randomness instead of rand()
The rand() function is a pseudo random number generator that always
supplies the same sequence of numbers unless reseeded.  Fix the fact
that sev-tool always uses the same random sequence by using the system
supplied getrandom() function instead.  This produces a source of
random numbers directly from the system entropy and is different on
each reboot.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>