SSL: Upgrade autoconf openssl api check
Adding in additional checks which guarantee that OpenSSL is greater than
version 1.1.0.x, as it is incompatible with sev-tool. Previously the
check for 1.1.1.x was looking for the API EVP_PKEY_base_id in the crypto
library. In the latest version 3.0.x of OpenSSL, that API was re-named
to EVP_PKEY_get_base_id. As both are acceptable version, we will check
for each of them.

Additional changes will need to be made before the next release to
address the deprecation warnings.

- Updated autoconf check for OpenSSL 3.0.x compatibility.

Signed-off-by: Larry Dewey <larry.dewey@amd.com>